remove public access level modifier from providers classes
