--- Sources/ThirdPartyAuth/Flows/Google/Handlers/GoogleSignInHandler.swift
@@ -1,18 +1,0 @@
-//
-//  GoogleSignInHandler.swift
-//  ThirdPartyAuth
-//
-//  Created by Ilya Klimenyuk on 01.05.2023.
-//
-
-//import GoogleSignIn
-//
-//public final class GoogleSignInHandler {
-//
-//    // MARK: - Public Methods
-//
-//    public func canHandle(_ url: URL) -> Bool {
-//        return GIDSignIn.sharedInstance.handle(url)
-//    }
-//
-//}
