--- Sources/ThirdPartyAuth/Flows/Google/Providers/GoogleAuthProvider.swift
@@ -1,94 +1,0 @@
-//
-//  GoogleAuthProvider.swift
-//  ThirdPartyAuth
-//
-//  Created by Ilya Klimenyuk on 01.05.2023.
-//
-
-import Foundation
-import GoogleSignIn
-
-/// Google sign in authorization process provider
-public final class GoogleAuthProvider: NSObject, BaseAuthProvider {
-
-    // MARK: - Public Properties
-
-    public var onAuthFinished: ((ThirdPartyAuthResult) -> Void)?
-
-    // MARK: - Public Methods
-
-    public func start(clientID: String) {
-        configureGoogleAuthInstance(with: clientID)
-    }
-
-    public func canHandle(_ url: URL) -> Bool {
-        return GIDSignIn.sharedInstance.handle(url)
-    }
-
-    public func restorePreviousSignIn() {
-        GIDSignIn.sharedInstance.restorePreviousSignIn { [weak self] user, error in
-            guard error == nil else {
-                self?.onAuthFinished?(.failure(.googleAuthFailed))
-                return
-            }
-
-            guard let user = user else {
-                self?.onAuthFinished?(.failure(.emptyGoogleUserData))
-                return
-            }
-
-            let userModel = ThirdPartyAuthUserModel(from: user)
-            self?.onAuthFinished?(.success(userModel))
-        }
-    }
-
-    public func performAuth() {
-        guard let topViewController = UIApplication.topViewController() else {
-            onAuthFinished?(.failure(.topViewControllerNotExist))
-            return
-        }
-
-        GIDSignIn.sharedInstance.signIn(withPresenting: topViewController) { [weak self] signInResult, error in
-            guard error == nil else {
-                self?.onAuthFinished?(.failure(.googleAuthFailed))
-                return
-            }
-
-            guard let signInResult = signInResult else {
-                self?.onAuthFinished?(.failure(.emptyGoogleUserData))
-                return
-            }
-
-            // Get user's ID token
-            signInResult.user.refreshTokensIfNeeded { [weak self] user, error in
-                guard error == nil else {
-                    self?.onAuthFinished?(.failure(.getGoogleUserIdTokenFailed))
-                    return
-                }
-
-                guard let user = user else {
-                    self?.onAuthFinished?(.failure(.emptyGoogleUserData))
-                    return
-                }
-
-                let userModel = ThirdPartyAuthUserModel(from: user)
-                self?.onAuthFinished?(.success(userModel))
-            }
-        }
-    }
-
-    public func signOut() {
-        GIDSignIn.sharedInstance.signOut()
-    }
-
-}
-
-// MARK: - Private Methods
-
-private extension GoogleAuthProvider {
-
-    func configureGoogleAuthInstance(with clientID: String) {
-        GIDSignIn.sharedInstance.configuration = .init(clientID: clientID)
-    }
-
-}
