+++ Sources/ThirdPartyAuth/Library/Protocols/BaseAuthProvider.swift
@@ -5,10 +5,11 @@
 //  Created by Ilya Klimenyuk on 21.04.2023.
 //
 
-public typealias ThirdPartyAuthResult = Result<ThirdPartyAuthUserModel, ThirdPartyAuthError>
+public typealias ThirdPartyAuthResult = Result<ThirdPartyAuthUserModel, Error>
 
 /// Protocol for third party authorization providers
-public protocol BaseAuthProvider {
+protocol BaseAuthProvider {
     var onAuthFinished: ((ThirdPartyAuthResult) -> Void)? { get set }
-    func performAuth()
+    func signIn()
+    func signOut()
 }
